float min, max and step props for numeric input

+++ src/View/Components/Fields/Input.php
@@ -27,9 +27,9 @@ class Input extends TwillFormComponent
         public ?string $onChange = null,
         public ?string $onChangeAttribute = null,
         public ?string $prefix = null,
-        public ?int $min = null,
-        public ?int $max = null,
-        public ?int $step = null,
+        public ?float $min = null,
+        public ?float $max = null,
+        public ?float $step = null,
         public ?string $mask = null,
     ) {
         parent::__construct(
